# docs/devicerepository_swagger.yaml
basePath: /
definitions:
...
  models.Function:
    properties:
      concept_id:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      name:
        type: string
      rdf_type:
        type: string
    type: object
...
  models.Location:
    properties:
      description:
        type: string
      device_group_ids:
        items:
          type: string
        type: array
      device_ids:
        items:
          type: string
        type: array
      id:
        type: string
      image:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Repository API
  version: "0.1"
paths:
...
  /functions:
    get:
      description: list functions
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter; https://senergy.infai.org/ontology/ControllingFunction
          || https://senergy.infai.org/ontology/MeasuringFunction
        in: query
        name: rdf_type
        type: string
      - description: filter
        in: query
        name: search
        type: string
      - description: default name.asc
        in: query
        name: sort
        type: string
      - description: filter; ignores limit/offset; comma-separated list
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: count of all matching elements; used for pagination
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.Function'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list functions
      tags:
      - functions
    post:
      description: create function
      parameters:
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Function'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Function'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create function
      tags:
      - functions
    put:
      consumes:
      - application/json
      description: validate location
      parameters:
      - description: must be true; reminder, that this is not an update but a validation
        in: query
        name: dry-run
        required: true
        type: boolean
      - description: Location to be validated
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: validate location
      tags:
      - locations
...
  /locations:
    get:
      description: list location
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter
        in: query
        name: search
        type: string
      - description: default name.asc
        in: query
        name: sort
        type: string
      - description: filter; ignores limit/offset; comma-separated list
        in: query
        name: ids
        type: string
      - description: default 'r'; used to check permissions on request; valid values
          are 'r', 'w', 'x', 'a' for read, write, execute, administrate
        in: query
        name: p
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: count of all matching elements; used for pagination
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.Location'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list location
      tags:
      - locations
    post:
      description: create location
      parameters:
      - description: location
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create location
      tags:
      - locations
  /locations/{id}:
    delete:
      description: delete location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete location
      tags:
      - locations
    get:
      description: get location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get location
      tags:
      - locations
    put:
      description: set location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set location
      tags:
      - locations
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"